+++ src/Console/ApplicationFactory.php
@@ -5,96 +5,189 @@ namespace Castor\Console;
 use Castor\Console\Command\CompileCommand;
 use Castor\Console\Command\DebugCommand;
 use Castor\Console\Command\RepackCommand;
-use Castor\ContextRegistry;
+use Castor\Container;
+use Castor\Event\AfterApplicationInitializationEvent;
 use Castor\EventDispatcher;
-use Castor\ExpressionLanguage;
-use Castor\Fingerprint\FingerprintHelper;
 use Castor\FunctionFinder;
 use Castor\Helper\PathHelper;
 use Castor\Helper\PlatformHelper;
-use Castor\Helper\WaitForHelper;
-use Castor\Import\Importer;
-use Castor\Import\Listener\RemoteImportListener;
-use Castor\Import\Remote\Composer;
-use Castor\Import\Remote\PackageImporter;
-use Castor\Listener\GenerateStubsListener;
-use Castor\Listener\UpdateCastorListener;
 use Castor\Monolog\Processor\ProcessProcessor;
-use Castor\Stub\StubsGenerator;
 use Monolog\Logger;
+use Psr\Cache\CacheItemPoolInterface;
+use Psr\Container\ContainerInterface;
+use Psr\Log\LoggerInterface;
+use Psr\Log\LogLevel;
 use Symfony\Component\Cache\Adapter\FilesystemAdapter;
+use Symfony\Component\Config\FileLocator;
 use Symfony\Component\Console\Application as SymfonyApplication;
+use Symfony\Component\Console\ConsoleEvents;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Style\SymfonyStyle;
+use Symfony\Component\DependencyInjection\ChildDefinition;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
+use Symfony\Component\DependencyInjection\Loader\PhpFileLoader;
+use Symfony\Component\ErrorHandler\ErrorHandler;
+use Symfony\Component\EventDispatcher\Attribute\AsEventListener;
+use Symfony\Component\EventDispatcher\DependencyInjection\RegisterListenersPass;
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\HttpClient\HttpClient;
+use Symfony\Component\VarDumper\Cloner\AbstractCloner;
+use Symfony\Contracts\Cache\CacheInterface;
+use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
+use Symfony\Contracts\HttpClient\HttpClientInterface;
+
+use function Symfony\Component\DependencyInjection\Loader\Configurator\service;
 
 /** @internal */
 class ApplicationFactory
 {
     public static function create(): SymfonyApplication
     {
+        AbstractCloner::$defaultCasters[self::class] = ['Symfony\Component\VarDumper\Caster\StubCaster', 'cutInternals'];
+        AbstractCloner::$defaultCasters[AfterApplicationInitializationEvent::class] = ['Symfony\Component\VarDumper\Caster\StubCaster', 'cutInternals'];
+
+        $handler = ErrorHandler::register();
+
         try {
             $rootDir = PathHelper::getRoot();
         } catch (\RuntimeException $e) {
             return new CastorFileNotFoundApplication($e);
         }
 
-        $class = Application::class;
-        if (class_exists(\RepackedApplication::class)) {
-            $class = \RepackedApplication::class;
-        }
-
-        $contextRegistry = new ContextRegistry();
-        $httpClient = HttpClient::create([
-            'headers' => [
-                'User-Agent' => 'Castor/' . Application::VERSION,
-            ],
+        $container = self::buildContainer();
+        $container->getParameterBag()->add([
+            'root_dir' => $rootDir,
+            'cache_dir' => PlatformHelper::getCacheDirectory(),
+            'event_dispatcher.event_aliases' => ConsoleEvents::ALIASES,
         ]);
-        $cacheDir = PlatformHelper::getCacheDirectory();
-        $cache = new FilesystemAdapter(directory: $cacheDir);
-        $logger = new Logger('castor', [], [new ProcessProcessor()]);
-        $fs = new Filesystem();
-        $fingerprintHelper = new FingerprintHelper($cache);
-        $packageImporter = new PackageImporter($logger, new Composer($fs, $logger, $fingerprintHelper));
-        $importer = new Importer($packageImporter, $logger);
-        $eventDispatcher = new EventDispatcher(logger: $logger);
-        $eventDispatcher->addSubscriber(new UpdateCastorListener(
-            $cache,
-            $httpClient,
-            $logger,
-        ));
-        $eventDispatcher->addSubscriber(new GenerateStubsListener(
-            new StubsGenerator($rootDir, $logger),
-        ));
-        $eventDispatcher->addSubscriber(new RemoteImportListener($packageImporter));
+        $container->compile();
 
-        /** @var Application */
+        $container->set(ContainerInterface::class, $container);
+
+        /** @var LoggerInterface */
+        $logger = $container->get(LoggerInterface::class);
+        $handler->setDefaultLogger($logger, [
+            \E_COMPILE_WARNING => LogLevel::WARNING,
+            \E_CORE_WARNING => LogLevel::WARNING,
+            \E_USER_WARNING => LogLevel::WARNING,
+            \E_WARNING => LogLevel::WARNING,
+            \E_USER_DEPRECATED => LogLevel::WARNING,
+            \E_DEPRECATED => LogLevel::WARNING,
+            \E_USER_NOTICE => LogLevel::WARNING,
+            \E_NOTICE => LogLevel::WARNING,
+
+            \E_COMPILE_ERROR => LogLevel::ERROR,
+            \E_CORE_ERROR => LogLevel::ERROR,
+            \E_ERROR => LogLevel::ERROR,
+            \E_PARSE => LogLevel::ERROR,
+            \E_RECOVERABLE_ERROR => LogLevel::ERROR,
+            \E_STRICT => LogLevel::ERROR,
+            \E_USER_ERROR => LogLevel::ERROR,
+        ]);
+
         // @phpstan-ignore-next-line
-        $application = new $class(
-            $rootDir,
-            new FunctionFinder($cache, $rootDir),
-            $contextRegistry,
-            $eventDispatcher,
-            new ExpressionLanguage($contextRegistry),
-            $logger,
-            $fs,
-            new WaitForHelper($httpClient, $logger),
-            $fingerprintHelper,
-            $importer,
-            $httpClient,
-            $cache,
-        );
+        return $container->get(Application::class);
+    }
 
-        // Avoid dependency cycle
-        $packageImporter->setApplication($application);
+    private static function buildContainer(): ContainerBuilder
+    {
+        $container = new ContainerBuilder();
 
-        $application->setDispatcher($eventDispatcher);
-        $application->add(new DebugCommand($rootDir, $cacheDir, $contextRegistry));
+        $container->registerForAutoconfiguration(EventSubscriberInterface::class)
+            ->addTag('kernel.event_subscriber')
+        ;
+        $container->addCompilerPass(new RegisterListenersPass());
+        // from https://github.com/symfony/symfony/blob/6.4/src/Symfony/Bundle/FrameworkBundle/DependencyInjection/FrameworkExtension.php#L676-L685
+        $container->registerAttributeForAutoconfiguration(AsEventListener::class, static function (ChildDefinition $definition, AsEventListener $attribute, \Reflector $reflector) {
+            $tagAttributes = get_object_vars($attribute);
+            if ($reflector instanceof \ReflectionMethod) {
+                if (isset($tagAttributes['method'])) {
+                    throw new \LogicException(sprintf('AsEventListener attribute cannot declare a method on "%s::%s()".', $reflector->class, $reflector->name));
+                }
+                $tagAttributes['method'] = $reflector->getName();
+            }
+            $definition->addTag('kernel.event_listener', $tagAttributes);
+        });
 
+        $phpLoader = new PhpFileLoader($container, new FileLocator());
+        $instanceof = [];
+        $configurator = new ContainerConfigurator($container, $phpLoader, $instanceof, __DIR__, __FILE__);
+        self::configureContainer($configurator);
+
+        return $container;
+    }
+
+    private static function configureContainer(ContainerConfigurator $c): void
+    {
+        $services = $c->services();
+
+        $services
+            ->defaults()
+                ->autowire()
+                ->autoconfigure()
+                ->bind('string $rootDir', '%root_dir%')
+                ->bind('string $cacheDir', '%cache_dir%')
+            ->load('Castor\\', __DIR__ . '/../*')
+            ->exclude([
+                __DIR__ . '/../functions.php',
+                __DIR__ . '/../functions-internal.php',
+            ])
+            ->set(FunctionFinder::class)
+            ->set(CacheInterface::class, FilesystemAdapter::class)
+                ->args([
+                    '$directory' => '%cache_dir%',
+                ])
+            ->alias(CacheItemPoolInterface::class . '&' . CacheInterface::class, CacheInterface::class)
+            ->set(HttpClientInterface::class)
+                ->factory([HttpClient::class, 'create'])
+                ->args([
+                    '$defaultOptions' => [
+                        'headers' => [
+                            'User-Agent' => 'Castor/' . Application::VERSION,
+                        ],
+                    ],
+                ])
+            ->set(Logger::class)
+                ->args([
+                    '$name' => 'castor',
+                    '$processors' => [
+                        service(ProcessProcessor::class),
+                    ],
+                ])
+            ->alias(LoggerInterface::class, Logger::class)
+                ->public()
+            ->alias(EventDispatcherInterface::class, EventDispatcher::class)
+            ->alias('event_dispatcher', EventDispatcherInterface::class)
+            ->set(Filesystem::class)
+            ->set(Container::class)
+                ->public()
+            ->set(ContainerInterface::class)
+                ->synthetic()
+            ->set(OutputInterface::class)
+                ->synthetic()
+                ->lazy()
+            ->set(InputInterface::class)
+                ->synthetic()
+            ->set(SymfonyStyle::class)
+        ;
+
+        $app = $services->set(Application::class, class_exists(\RepackedApplication::class) ? \RepackedApplication::class : null)
+                ->public()
+                ->args([
+                    '$containerBuilder' => service(ContainerInterface::class),
+                ])
+                ->call('add', [service(DebugCommand::class)])
+                ->call('setDispatcher', [service(EventDispatcherInterface::class)])
+                ->call('setCatchErrors', [true])
+        ;
         if (!class_exists(\RepackedApplication::class)) {
-            $application->add(new RepackCommand());
-            $application->add(new CompileCommand($httpClient, $fs));
+            $app
+                ->call('add', [service(RepackCommand::class)])
+                ->call('add', [service(CompileCommand::class)])
+            ;
         }
-
-        return $application;
     }
 }